requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
